# Perf: Optimize AssemblyReference.GetAssembly