--- a/src/Perf/Perf.AssemblyReference.Benchmarks/Perf.AssemblyReference.Benchmarks.csproj
+++ b/src/Perf/Perf.AssemblyReference.Benchmarks/Perf.AssemblyReference.Benchmarks.csproj
@@ -1,0 +1,40 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+	<PropertyGroup>
+		<OutputType>Exe</OutputType>
+		<TargetFrameworks>net6.0;net6.0-windows</TargetFrameworks>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<Nullable>enable</Nullable>
+	</PropertyGroup>
+
+	<ItemGroup>
+		<PackageReference Include="AutoMapper" Version="9.0.0" />
+		<PackageReference Include="BenchmarkDotNet" Version="0.13.1" />
+		<PackageReference Include="Bogus" Version="35.6.3" />
+		<PackageReference Include="ClosedXML" Version="0.105.0" />
+		<PackageReference Include="CsvHelper" Version="33.1.0" />
+		<PackageReference Include="Dapper" Version="2.1.66" />
+		<PackageReference Include="EPPlus" Version="8.0.6" />
+		<PackageReference Include="FluentValidation" Version="6.4.1" />
+		<PackageReference Include="Hangfire.Core" Version="1.8.20" />
+		<PackageReference Include="HtmlAgilityPack" Version="1.12.1" />
+		<PackageReference Include="Humanizer" Version="2.14.1" />
+		<PackageReference Include="MailKit" Version="4.12.1" />
+		<PackageReference Include="MediatR" Version="12.5.0" />
+		<PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
+		<PackageReference Include="NLog" Version="6.0.0" />
+		<PackageReference Include="NodaTime" Version="3.2.2" />
+		<PackageReference Include="Polly" Version="8.6.1" />
+		<PackageReference Include="RestSharp" Version="112.1.0" />
+		<PackageReference Include="Serilog" Version="4.3.0" />
+		<PackageReference Include="Serilog.Sinks.Console" Version="6.0.0" />
+		<PackageReference Include="SixLabors.ImageSharp" Version="3.1.10" />
+		<PackageReference Include="System.Drawing.Common" Version="6.0.0" />
+	</ItemGroup>
+
+	<ItemGroup>
+		<ProjectReference Include="..\..\UiPath.Workflow.Runtime\UiPath.Workflow.Runtime.csproj" />
+		<ProjectReference Include="..\..\UiPath.Workflow\UiPath.Workflow.csproj" />
+	</ItemGroup>
+
+</Project>