--- a/src/Perf/Perf.AssemblyReference.Benchmarks/Perf.AssemblyReference.Benchmarks.csproj
+++ b/src/Perf/Perf.AssemblyReference.Benchmarks/Perf.AssemblyReference.Benchmarks.csproj
@@ -9,6 +9,9 @@
 
 	<ItemGroup>
 		<PackageReference Include="AutoMapper" Version="9.0.0" />
+		<PackageReference Include="Azure.Identity" Version="1.14.0" />
+		<PackageReference Include="Azure.Messaging.ServiceBus" Version="7.20.1" />
+		<PackageReference Include="Azure.Storage.Blobs" Version="12.24.1" />
 		<PackageReference Include="BenchmarkDotNet" Version="0.13.1" />
 		<PackageReference Include="Bogus" Version="35.6.3" />
 		<PackageReference Include="ClosedXML" Version="0.105.0" />
@@ -16,20 +19,46 @@
 		<PackageReference Include="Dapper" Version="2.1.66" />
 		<PackageReference Include="EPPlus" Version="8.0.6" />
 		<PackageReference Include="FluentValidation" Version="6.4.1" />
+		<PackageReference Include="Google.Apis" Version="1.70.0" />
+		<PackageReference Include="Google.Apis.Auth" Version="1.70.0" />
+		<PackageReference Include="Google.Cloud.Storage.V1" Version="4.13.0" />
 		<PackageReference Include="Hangfire.Core" Version="1.8.20" />
 		<PackageReference Include="HtmlAgilityPack" Version="1.12.1" />
 		<PackageReference Include="Humanizer" Version="2.14.1" />
 		<PackageReference Include="MailKit" Version="4.12.1" />
 		<PackageReference Include="MediatR" Version="12.5.0" />
+		<PackageReference Include="Microsoft.AspNetCore" Version="2.3.0" />
+		<PackageReference Include="Microsoft.AspNetCore.Authentication" Version="2.3.0" />
+		<PackageReference Include="Microsoft.AspNetCore.Authorization" Version="9.0.6" />
+		<PackageReference Include="Microsoft.AspNetCore.Http" Version="2.3.0" />
+		<PackageReference Include="Microsoft.AspNetCore.Mvc" Version="2.3.0" />
+		<PackageReference Include="Microsoft.AspNetCore.Routing" Version="2.3.0" />
+		<PackageReference Include="Microsoft.AspNetCore.StaticFiles" Version="2.3.0" />
+		<PackageReference Include="Microsoft.Azure.Cosmos" Version="3.52.0" />
+		<PackageReference Include="Microsoft.Azure.Functions.Extensions" Version="1.1.0" />
+		<PackageReference Include="Microsoft.Azure.ServiceBus" Version="5.2.0" />
+		<PackageReference Include="Microsoft.Azure.Storage.Blob" Version="11.2.3" />
+		<PackageReference Include="Microsoft.Extensions.Configuration" Version="9.0.6" />
+		<PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="9.0.6" />
+		<PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="9.0.6" />
+		<PackageReference Include="Microsoft.Extensions.Logging" Version="9.0.6" />
+		<PackageReference Include="Microsoft.Extensions.Options" Version="9.0.6" />
+		<PackageReference Include="Microsoft.IdentityModel.Protocols" Version="8.12.1" />
+		<PackageReference Include="Microsoft.IdentityModel.Tokens" Version="8.12.1" />
+		<PackageReference Include="MongoDB.Driver" Version="3.4.0" />
 		<PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
 		<PackageReference Include="NLog" Version="6.0.0" />
 		<PackageReference Include="NodaTime" Version="3.2.2" />
+		<PackageReference Include="Npgsql" Version="9.0.3" />
 		<PackageReference Include="Polly" Version="8.6.1" />
 		<PackageReference Include="RestSharp" Version="112.1.0" />
 		<PackageReference Include="Serilog" Version="4.3.0" />
 		<PackageReference Include="Serilog.Sinks.Console" Version="6.0.0" />
 		<PackageReference Include="SixLabors.ImageSharp" Version="3.1.10" />
+		<PackageReference Include="System.Data.SqlClient" Version="4.9.0" />
 		<PackageReference Include="System.Drawing.Common" Version="6.0.0" />
+		<PackageReference Include="System.IO.Pipelines" Version="9.0.6" />
+		<PackageReference Include="System.Reactive" Version="6.0.1" />
 	</ItemGroup>
 
 	<ItemGroup>