--- a/src/Perf/Perf.AssemblyReference.Benchmarks/Program.cs
+++ b/src/Perf/Perf.AssemblyReference.Benchmarks/Program.cs
@@ -1,0 +1,4 @@
+﻿using BenchmarkDotNet.Running;
+using Perf.AssemblyReference.Benchmarks;
+
+BenchmarkRunner.Run<AssemblyReferenceBenchmarks>();