--- a/src/UiPath.Workflow.Runtime/AssemblyInfo.cs
+++ b/src/UiPath.Workflow.Runtime/AssemblyInfo.cs
@@ -28,3 +28,4 @@ using System.Windows.Markup;
 [assembly: InternalsVisibleTo("UiPath.Workflow")]
 [assembly: InternalsVisibleTo("TestCases.Workflows")]
 [assembly: InternalsVisibleTo("TestCases.Runtime")]
+[assembly: InternalsVisibleTo("Perf.AssemblyReference.Benchmarks")]