Benchmark below is for 10.000 lookups, and 377 assemblies in the load context:

```
|                             Method |            Mean |         Error |        StdDev |     Gen 0 |     Gen 1 |     Allocated |
|----------------------------------- |----------------:|--------------:|--------------:|----------:|----------:|--------------:|
|                   SequentialLookup |     2,023.28 us |     30.129 us |     28.182 us |   23.4375 |         - |     410,597 B |
|         SequentialLookup_CacheMiss |   534,399.22 us | 10,622.933 us | 24,407.955 us | 7000.0000 |         - | 129,361,632 B |
|               ParallelLookup4Cores |       100.38 us |      1.969 us |      4.197 us |    0.1221 |         - |       2,312 B |
|              ParallelLookup16Cores |        58.10 us |      0.882 us |      0.782 us |    0.3052 |         - |       5,287 B |
|    ParallelLookup16Cores_CacheMiss | 2,335,725.04 us | 38,796.851 us | 36,290.598 us | 8000.0000 | 1000.0000 | 129,372,080 B |
|                SequentialLookup_V2 |       483.47 us |      9.061 us |      7.566 us |         - |         - |           1 B |
|      SequentialLookup_CacheMiss_V2 |       335.30 us |      4.567 us |      4.049 us |         - |         - |           1 B |
|            ParallelLookup4Cores_V2 |       103.38 us |      1.926 us |      1.801 us |    0.1221 |         - |       2,312 B |
|           ParallelLookup16Cores_V2 |        60.61 us |      1.208 us |      1.612 us |    0.3052 |         - |       5,288 B |
| ParallelLookup16Cores_CacheMiss_V2 |       900.95 us |     11.278 us |      9.418 us |         - |         - |       5,291 B |
```